+++ src/steps/account/converter.ts
@@ -5,21 +5,17 @@ import {
 
 import { Entities } from '../constants';
 
-export function createAccountEntity(): Entity {
+export function createAccountEntity(organization: string): Entity {
   return createIntegrationEntity({
     entityData: {
       source: {
-        id: 'acme-unique-account-id',
-        name: 'Example Co. Acme Account',
+        id: organization,
+        name: 'Polymer Account',
       },
       assign: {
-        _key: 'acme-unique-account-id',
+        _key: `${Entities.ACCOUNT._type}|organization`,
         _type: Entities.ACCOUNT._type,
         _class: Entities.ACCOUNT._class,
-        mfaEnabled: true,
-        // This is a custom property that is not a part of the data model class
-        // hierarchy. See: https://github.com/JupiterOne/data-model/blob/master/src/schemas/Account.json
-        manager: 'Manager Name',
       },
     },
   });