+++ src/steps/account/index.ts
@@ -10,9 +10,12 @@ import { createAccountEntity } from './converter';
 export const ACCOUNT_ENTITY_KEY = 'entity:account';
 
 export async function fetchAccountDetails({
+  instance,
   jobState,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
-  const accountEntity = await jobState.addEntity(createAccountEntity());
+  const accountEntity = await jobState.addEntity(
+    createAccountEntity(instance.config.organization),
+  );
 
   await jobState.setData(ACCOUNT_ENTITY_KEY, accountEntity);
 }