+++ src/steps/violations/index.test.ts
@@ -3,26 +3,19 @@ import { buildStepTestConfigForStep } from '../../../test/config';
 import { Recording, setupProjectRecording } from '../../../test/recording';
 import { Steps } from '../constants';
 
-function isRecordingEnabled() {
-  return Boolean(process.env.LOAD_ENV) === true;
-}
-
 // See test/README.md for details
 let recording: Recording;
 afterEach(async () => {
   await recording.stop();
 });
 
-test('fetch-account', async () => {
+test('fetch-violations', async () => {
   recording = setupProjectRecording({
     directory: __dirname,
-    name: 'fetch-account',
-    options: {
-      mode: isRecordingEnabled() ? 'record' : 'replay',
-    },
+    name: 'fetch-violations',
   });
 
-  const stepConfig = buildStepTestConfigForStep(Steps.ACCOUNT);
+  const stepConfig = buildStepTestConfigForStep(Steps.VIOLATIONS);
   const stepResult = await executeStepWithDependencies(stepConfig);
   expect(stepResult).toMatchStepMetadata(stepConfig);
 });