+++ src/validateInvocation.test.ts
@@ -1,3 +1,4 @@
+import { IntegrationProviderAuthenticationError } from '@jupiterone/integration-sdk-core';
 import {
   createMockExecutionContext,
   Recording,
@@ -21,14 +22,14 @@ describe('#validateInvocation', () => {
     });
 
     await expect(validateInvocation(executionContext)).rejects.toThrow(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires all of {apiToken, organization}',
     );
   });
 
   /**
    * Testing a successful authorization can be done with recordings
    */
-  test.skip('successfully validates invocation', async () => {
+  test('successfully validates invocation', async () => {
     recording = setupProjectRecording({
       directory: __dirname,
       name: 'validate-invocation',
@@ -43,10 +44,6 @@ describe('#validateInvocation', () => {
     await expect(validateInvocation(executionContext)).resolves.toBeUndefined();
   });
 
-  /* Adding `describe` blocks segments the tests into logical sections
-   * and makes the output of `yarn test --verbose` provide meaningful
-   * to project information to future maintainers.
-   */
   describe('fails validating invocation', () => {
     /**
      * Testing failing authorizations can be done with recordings as well.
@@ -54,7 +51,7 @@ describe('#validateInvocation', () => {
      * error messaging is expected and clear to end-users
      */
     describe('invalid user credentials', () => {
-      test.skip('should throw if clientId is invalid', async () => {
+      test('should throw if login is invalid', async () => {
         recording = setupProjectRecording({
           directory: __dirname,
           name: 'client-id-auth-error',
@@ -67,36 +64,15 @@ describe('#validateInvocation', () => {
 
         const executionContext = createMockExecutionContext({
           instanceConfig: {
-            clientId: 'INVALID',
-            clientSecret: integrationConfig.clientSecret,
+            apiToken: 'INVALID',
+            organization: integrationConfig.organization,
           },
         });
 
         // tests validate that invalid configurations throw an error
         // with an appropriate and expected message.
         await expect(validateInvocation(executionContext)).rejects.toThrow(
-          'Provider authentication failed at https://localhost/api/v1/some/endpoint?limit=1: 401 Unauthorized',
-        );
-      });
-
-      test.skip('should throw if clientSecret is invalid', async () => {
-        recording = setupProjectRecording({
-          directory: __dirname,
-          name: 'client-secret-auth-error',
-          options: {
-            recordFailedRequests: true,
-          },
-        });
-
-        const executionContext = createMockExecutionContext({
-          instanceConfig: {
-            clientId: integrationConfig.clientSecret,
-            clientSecret: 'INVALID',
-          },
-        });
-
-        await expect(validateInvocation(executionContext)).rejects.toThrow(
-          'Provider authentication failed at https://localhost/api/v1/some/endpoint?limit=1: 401 Unauthorized',
+          IntegrationProviderAuthenticationError,
         );
       });
     });